Refactored tests utils
