+++ ydb/core/fq/libs/row_dispatcher/format_handler/ut/common/ya.make
@@ -15,6 +15,7 @@ PEERDIR(
     ydb/core/testlib/actors
     ydb/core/testlib/basics
 
+    ydb/library/testlib/common
     ydb/library/yql/dq/actors
     ydb/library/yql/dq/common
