+++ ydb/public/sdk/cpp/src/client/topic/ut/ut_utils/topic_sdk_test_setup.cpp
@@ -1,51 +1,17 @@
 #include "topic_sdk_test_setup.h"
 
 #include <ydb/core/base/backtrace.h>
+#include <ydb/library/testlib/common/test_utils.h>
 
 using namespace std::chrono_literals;
 
 namespace NYdb::inline Dev::NTopic::NTests {
 
-namespace {
-
-void TerminateHandler() {
-    Cerr << "======= terminate() call stack ========" << Endl;
-    FormatBackTrace(&Cerr);
-    if (const auto& backtrace = TBackTrace::FromCurrentException(); backtrace.size() > 0) {
-        Cerr << "======== exception call stack =========" << Endl;
-        backtrace.PrintTo(Cerr);
-    }
-    Cerr << "=======================================" << Endl;
-
-    if (std::current_exception()) {
-        Cerr << "Uncaught exception: " << CurrentExceptionMessage() << Endl;
-    } else {
-        Cerr << "Terminate for unknown reason (no current exception)" << Endl;
-    }
-
-    abort();
-}
-
-void BackTraceSignalHandler(int signal) {
-    Cerr << "======= Signal " << signal << " call stack ========" << Endl;
-    FormatBackTrace(&Cerr);
-    Cerr << "===============================================" << Endl;
-
-    abort();
-}
-
-} // anonymous namespace
-
 TTopicSdkTestSetup::TTopicSdkTestSetup(const std::string& testCaseName, const NKikimr::Tests::TServerSettings& settings, bool createTopic)
     : Database_("/Root")
     , Server_(settings, false)
 {
-    NKikimr::EnableYDBBacktraceFormat();
-
-    std::set_terminate(&TerminateHandler);
-    for (auto sig : {SIGFPE, SIGILL, SIGSEGV}) {
-        signal(sig, &BackTraceSignalHandler);
-    }
+    NTestUtils::SetupSignalHandlers();
 
     Log_.SetFormatter([testCaseName](ELogPriority priority, TStringBuf message) {
         return TStringBuilder() << TInstant::Now() << " :" << testCaseName << " " << priority << ": " << message << Endl;
