+++ ydb/tests/fq/control_plane_storage/in_memory_control_plane_storage_ut.cpp
@@ -4,6 +4,7 @@
 
 #include <ydb/library/actors/core/executor_pool_basic.h>
 #include <ydb/library/actors/core/scheduler_basic.h>
+#include <ydb/library/testlib/common/test_utils.h>
 #include <library/cpp/testing/unittest/registar.h>
 
 #include <ydb/core/testlib/basics/runtime.h>
@@ -35,8 +36,7 @@ public:
     void SetUp(NUnitTest::TTestContext& ctx) override {
         TBase::SetUp(ctx);
 
-        SetupSignalActions();
-        EnableYDBBacktraceFormat();
+        NTestUtils::SetupSignalHandlers();
 
         TFqSetupSettings settings;
         settings.VerbosityLevel = TFqSetupSettings::EVerbosity::InitLogs;
