+++ ydb/tests/tools/kqprun/kqprun.cpp
@@ -10,6 +10,7 @@
 #include <ydb/core/blob_depot/mon_main.h>
 #include <ydb/core/protos/table_service_config.pb.h>
 #include <ydb/library/aclib/aclib.h>
+#include <ydb/library/testlib/common/test_utils.h>
 #include <ydb/library/yaml_config/yaml_config.h>
 #include <ydb/library/yql/providers/pq/gateway/dummy/yql_pq_dummy_gateway.h>
 #include <ydb/tests/tools/kqprun/runlib/application.h>
@@ -1138,7 +1139,7 @@ private:
 }  // namespace NKqpRun
 
 int main(int argc, const char* argv[]) {
-    SetupSignalActions();
+    NTestUtils::SetupSignalHandlers();
 
 #ifdef PROFILE_MEMORY_ALLOCATIONS
     NMonitoring::TDynamicCounterPtr memoryProfilingCounters = MakeIntrusive<NMonitoring::TDynamicCounters>();
